chore: add the collides? method to the racket
Add a new method to the Racket class to detect if an entity collides with it.
The Racket rect is transformed with its shape's local bounds since the shape is
what is used to check for intersections.

-------------------------------------######-------------------------------------

The local geometric of the entity is the rectangle shape, since it's what
internally composes the Racket entity, so the Racket's local bounding rect is
just the shape's local bounds¹.

Reference:
¹https://en.sfml-dev.org/forums/index.php?topic=22309.msg157114#msg157114